requirements.txt
Django==4.1.2
Flask==2.2.2
gunicorn==20.1.0
whitenoise==6.2.0
pytesseract
matplotlib
numpy
opencv-python-headless
Pillow